requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
requests
streamlit
